requirements.txt
torch==2.2.2
typer==0.15.1
fastapi==0.115.6
uvicorn==0.34.0
pytorch-lightning==2.5.0
hydra-core==1.3.2
transformers==4.47.1
wandb==0.19.2
numpy==1.26.4
scikit-learn==1.5.2
pandas==2.2.3
loguru==0.7.3
python-dotenv
